Fought against authorisation issues and looping Postman requests due to infinite references, settled on using JSON ignore to stop issue.

+++ src/main/java/com/example/SongTracker_Backend/Exceptions/AlbumNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.SongTracker_Backend.Exceptions;
 
 public class AlbumNotFoundException extends RuntimeException {
-    public AlbumNotFoundException(long album_id) {
+    public AlbumNotFoundException(Integer album_id) {
         super("Could not find artist " + album_id);
     }
 }
