+++ src/main/java/com/example/SongTracker_Backend/Exceptions/ArtistNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.SongTracker_Backend.Exceptions;
 
 public class ArtistNotFoundException extends RuntimeException {
-    public ArtistNotFoundException(long artist_id){
+    public ArtistNotFoundException(Integer artist_id){
         super("Could not find artist " + artist_id);
     }
 }
