+++ src/main/java/com/example/SongTracker_Backend/Exceptions/SongNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.SongTracker_Backend.Exceptions;
 
 public class SongNotFoundException extends RuntimeException{
-    public SongNotFoundException(long song_id){
+    public SongNotFoundException(Integer song_id){
         super("Could not find song " + song_id);
     }
 }
